Add missing package exports. Extract interface from entity so not to leak the internal type.
