+++ libs/scheduler/scheduler-datamodel/src/main/java/net/corda/libs/scheduler/datamodel/db/internal/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.libs.scheduler.datamodel.db.internal;
+
+import org.osgi.annotation.bundle.Export;
