+++ libs/scheduler/scheduler-datamodel/src/main/java/net/corda/libs/scheduler/datamodel/db/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.libs.scheduler.datamodel.db;
+
+import org.osgi.annotation.bundle.Export;
