+++ libs/scheduler/scheduler-datamodel/src/main/kotlin/net/corda/libs/scheduler/datamodel/db/TaskSchedulerLog.kt
@@ -1,0 +1,10 @@
+package net.corda.libs.scheduler.datamodel.db
+
+import java.time.Instant
+
+interface TaskSchedulerLog {
+    var name: String
+    var schedulerId: String
+    var lastScheduled: Instant
+    var now: Instant
+}
