+++ libs/scheduler/scheduler-datamodel/src/main/kotlin/net/corda/libs/scheduler/datamodel/db/TaskSchedulerLogEntityRepository.kt
@@ -1,5 +1,10 @@
-package net.corda.libs.scheduler.datamodel.db.internal
+package net.corda.libs.scheduler.datamodel.db
 
+import net.corda.libs.scheduler.datamodel.db.internal.TASK_SCHEDULER_LOG_GET_QUERY_NAME
+import net.corda.libs.scheduler.datamodel.db.internal.TASK_SCHEDULER_LOG_QUERY_PARAM_NAME
+import net.corda.libs.scheduler.datamodel.db.internal.TASK_SCHEDULER_LOG_QUERY_PARAM_SCHEDULER_ID
+import net.corda.libs.scheduler.datamodel.db.internal.TASK_SCHEDULER_LOG_UPDATE_QUERY_NAME
+import net.corda.libs.scheduler.datamodel.db.internal.TaskSchedulerLogEntity
 import java.time.Instant
 import java.util.Date
 import javax.persistence.EntityManager
@@ -12,11 +17,11 @@ class TaskSchedulerLogEntityRepository {
     /**
      * Get the latest log for a given `taskName`, or initialise a new one if one doesn't exist.
      */
-    fun getOrInitialiseLog(taskName: String, schedulerId: String, em: EntityManager) : TaskSchedulerLogEntity {
+    fun getOrInitialiseLog(taskName: String, schedulerId: String, em: EntityManager) : TaskSchedulerLog {
         val readQuery = em.createNamedQuery(TASK_SCHEDULER_LOG_GET_QUERY_NAME, TaskSchedulerLogEntity::class.java)
         readQuery.setParameter(TASK_SCHEDULER_LOG_QUERY_PARAM_NAME, taskName)
         readQuery.lockMode = LockModeType.PESSIMISTIC_WRITE
-        return readQuery.resultList.singleOrNull()?:TaskSchedulerLogEntity(taskName, schedulerId, Instant.MIN, Date.from(
+        return readQuery.resultList.singleOrNull()?: TaskSchedulerLogEntity(taskName, schedulerId, Instant.MIN, Date.from(
             Instant.now()))
     }
 
@@ -30,3 +35,4 @@ class TaskSchedulerLogEntityRepository {
         updateQ.executeUpdate()
     }
 }
+
